Added course sync
