+++ .vscode/launch.json
@@ -30,6 +30,8 @@
                 "${workspaceFolder}/scratch/data-samples/ksPERMRECS.xml.json",
                 "--teacher-file",
                 "${workspaceFolder}/scratch/data-samples/ksTEACHERS.xml.json",
+                "--course-file",
+                "${workspaceFolder}/scratch/data-samples/ksCOURSES.xml.json",
             ],
         }
     ]
